+++ react/components/AccordionFilterGroup.js
@@ -24,6 +24,8 @@ const AccordionFilterGroup = ({
   navigationType,
   initiallyCollapsed,
   truncateFilters,
+  onClearFilter,
+  showClearByFilter,
 }) => {
   const handles = useCssHandles(CSS_HANDLES)
   const filters = useSelectedFilters(facets)
@@ -31,9 +33,11 @@ const AccordionFilterGroup = ({
   const intl = useIntl()
   const facetTitle = getFilterTitle(title, intl)
   const slugifiedFacetTitle = searchSlugify(facetTitle)
+  const facetKey = filters.length > 0 ? filters[0].key : null
 
   return (
     <AccordionFilterItem
+      facetKey={facetKey}
       title={title}
       open={open}
       show={show}
@@ -42,6 +46,9 @@ const AccordionFilterGroup = ({
       appliedFiltersOverview={appliedFiltersOverview}
       navigationType={navigationType}
       initiallyCollapsed={initiallyCollapsed}
+      onFilterCheck={onFilterCheck}
+      onClearFilter={onClearFilter}
+      showClearByFilter={showClearByFilter}
     >
       <div
         className={classNames(